chore: pre-commit linting (#54)
* chore: script to lint mdx files by parsing it

To implement auto translation, we need to create a script to
separate some nodes such as import statement and code block using
`remark` package. For that script to work, the mdx files need
to be parseable by `remark`. Therefore, we need to check before commit
whether all mdx files are parseable.

* fix: fix mdx syntax errors

There are 4 common errors:
1. Passing a JSX element as prop without curly braces.
2. `<a>` tag not started in new line
3. Whitespace expression `{" "}` right after `</a>` on the same line.
4. Whitespaces before JSX component, import statement, paragraph, etc.
   It can be fixed by using prettier.

Also, I remove unused `alt` prop that's passed to `IfcImage` component.

* chore: run yarn lint before every commit

* add prettier

* chore: format using prettier

By formatting `.mdx` files with prettier, it will minimize mdx
parsing error that being caught by the linter.

This commit also fix:
1. Syntax error after being formatted
2. Non-syntax error that make the formatting unstable (previous
   to next formatting have different result when there is no change).

TODO:
1. Create MDX authoring guidelines to prevent and fix syntax errors
   and formatting unstability.

* chore: runt format linter before every commit

* Provide common solutions for MDX parsing errors and volatile formatting

* format with prettier

* format staged files instead of --list-different

Turns out `prettier --list-different` will always erroring out in
Windows if the end of line character is converted from Linux style to
Windows style by git when executing git checkout, pull, etc.

So, I think it's better to automatically format staged files before
every commit.

* remove formatting first guideline in MDX.md

* change prettier printWidth for .md and .mdx files.

The whitespace `{" "}` JavaScript expression will be treated as
text by Deepl. Sometimes, it will be relocated (e.g. in Japanese)
from the end of line to the middle of line which is not desired.

By setting the printWidth to Infinity, the text will not be wrapped
to next line, so there is no need for the whitespace `{" "}` expression.